Shock Pots
Added Shock Pots to the code, also fixed the data row shifting issue by
making MOTEC the master timer for the system.  Whatever frequency MOTEC
transmits data at is the frequency at which the system logs at.  This
may be an issue when it comes time to implement multi frequency logging.